refactor: extract timetable update logic into `TimetableUpdateMapper`
